presentation folder reorganization, scenemanager clean up
