+++ src/main/java/presentation/view/SceneManager.java
@@ -1,43 +1,32 @@
 package presentation.view;
+
 import domain.model.StudyMaterial;
 import domain.model.User;
-import domain.service.RatingService;
-import infrastructure.repository.RatingRepository;
+import domain.model.Category;
+
 import infrastructure.repository.StudyMaterialRepository;
 import infrastructure.repository.UserRepository;
-import javafx.geometry.Pos;
-import javafx.scene.control.*;
-import javafx.scene.image.Image;
-import javafx.scene.image.ImageView;
-import javafx.scene.text.TextFlow;
-import presentation.components.MaterialCard;
-import presentation.components.Stars;
-import presentation.components.TextLabels;
-
-import domain.model.Category;
 import infrastructure.repository.CategoryRepository;
+
+import javafx.scene.Node;
+import javafx.scene.control.*;
+import javafx.scene.layout.*;
 
 import javafx.fxml.FXMLLoader;
 import javafx.geometry.Insets;
 import javafx.scene.Scene;
-import javafx.scene.layout.BorderPane;
-import javafx.scene.layout.GridPane;
-import javafx.scene.layout.HBox;
-import javafx.scene.layout.VBox;
 import javafx.scene.text.Text;
 import javafx.stage.Stage;
 
-import java.io.ByteArrayInputStream;
 import java.io.IOException;
-import java.time.format.DateTimeFormatter;
-import java.util.ArrayList;
-import java.util.List;
 import java.util.Objects;
 
-import static presentation.components.ListItem.listItemGraphic;
 import static presentation.view.Screen.*;
-import presentation.components.ListItem;
-import presentation.logger.GUILogger;
+
+import presentation.components.CategoryPage;
+import presentation.components.MaterialPage;
+import presentation.components.ProfilePage;
+import presentation.GUILogger;
 
 public class SceneManager {
     private static SceneManager instance;
@@ -80,45 +69,7 @@ public class SceneManager {
                 GUILogger.warn("DNE: Tried to go to category with id " + id);
                 displayErrorPage("This category does not exist.", SCREEN_HOME, "Go to home page");
             } else {
-                VBox vbox = new VBox();
-
-                vbox.getStylesheets().add(Objects.requireNonNull(getClass().getResource("/css/style.css")).toExternalForm());
-                vbox.setSpacing(12);
-                vbox.setPadding(new Insets(20, 20, 20, 20));
-                Text title = new Text(c.getCategoryName());
-                title.getStyleClass().add("heading3");
-                title.getStyleClass().add("secondary");
-
-                Text author = new Text("Course by " + c.getCreator().getFullName());
-                VBox header = new VBox();
-                header.getChildren().addAll(title, author);
-
-                vbox.getChildren().add(header);
-
-                List<StudyMaterial> creatorMaterials = repo.findMaterialsByUserInCategory(c.getCreator(), c);
-                if (!creatorMaterials.isEmpty()) {
-                    Text text = new Text("Materials from " + c.getCreator().getFullName());
-                    text.getStyleClass().add("heading4");
-                    text.getStyleClass().add("secondary");
-
-                    vbox.getChildren().addAll(
-                            text,
-                            MaterialCard.materialCardScrollHBox(creatorMaterials));
-                }
-
-                List<StudyMaterial> otherMaterials = repo.findMaterialsExceptUserInCategory(c.getCreator(), c);
-                if (!otherMaterials.isEmpty()) {
-                    GUILogger.info(String.valueOf(otherMaterials.size()));
-                    Text text = new Text("Materials from others");
-                    text.getStyleClass().add("heading4");
-                    text.getStyleClass().add("secondary");
-
-                    vbox.getChildren().addAll(
-                            text,
-                            MaterialCard.materialCardScrollHBox(otherMaterials));
-                }
-
-                instance.current.setCenter(vbox);
+                CategoryPage.setPage(c);
             }
         }
     }
@@ -138,98 +89,7 @@ public class SceneManager {
             return;
         }
 
-        /* header: preview and file details*/
-        VBox base = new VBox();
-        base.getStylesheets().add(Objects.requireNonNull(getClass().getResource("/css/style.css")).toExternalForm());
-        base.setSpacing(12);
-        base.setPadding(new Insets(20, 20, 20, 20));
-
-        HBox main = new HBox();
-
-        VBox left = new VBox();
-        Label title = new Label(s.getName());
-        title.getStyleClass().add("label3");
-        title.getStyleClass().add("primary-light");
-        title.setWrapText(true);
-        title.setMaxWidth(600);
-
-        TextFlow uploaderLabels = new TextFlow();
-        Text author = new Text("Uploaded by ");
-        author.setStyle("-fx-font-size: 1.2em;");
-
-        Hyperlink authorLink = new Hyperlink(s.getUploader().getFullName());
-        authorLink.setStyle("-fx-font-size: 1.2em; -fx-underline: false;");
-        authorLink.setOnAction(e -> {
-            try {
-                displayProfile(s.getUploader().getUserId());
-            } catch (IOException ex) {
-                displayErrorPage("User not found.", SCREEN_HOME, "Go to Home");
-            }
-        });
-
-        uploaderLabels.getChildren().addAll(author, authorLink, new Text("  "), TextLabels.getUserRoleLabel(s.getUploader()), new Text("  "));
-
-        if (s.getUploader() == s.getCategory().getCreator()) {
-            Label categoryOwnerLabel = new Label("Course Owner");
-            categoryOwnerLabel.getStyleClass().add("primaryTagLabel");
-            uploaderLabels.getChildren().add(categoryOwnerLabel);
-        }
-
-        Text fileDetails = new Text(Math.round(s.getFileSize()) + " KB " + s.getFileType());
-
-        Button downloadBtn = new Button("Download");
-        downloadBtn.getStyleClass().add("btnDownload");
-        downloadBtn.setOnAction(event -> GUILogger.info("Pressed button to download " + s.getName()));
-
-        TextFlow fileDesc = new TextFlow();
-        fileDesc.getChildren().add(new Text(s.getDescription()));
-        fileDesc.setMaxWidth(580);
-
-        DateTimeFormatter formatter = DateTimeFormatter.ofPattern("dd.MM.yy HH:mm");
-        String formattedTimestamp = s.getTimestamp().format(formatter);
-
-        Text course = new Text("Uploaded under course " + s.getCategory().getCategoryName() + " on " + formattedTimestamp);
-        course.getStyleClass().add("primary");
-
-        left.getChildren().addAll(title, uploaderLabels, fileDetails, downloadBtn,
-                                    course, fileDesc);
-        left.setMinWidth(580);
-        left.setMaxWidth(580);
-        left.setSpacing(8);
-
-        VBox right = new VBox();
-        ImageView preview = new ImageView(new Image(new ByteArrayInputStream(s.getPreviewImage())));
-        preview.setFitWidth(141);
-        preview.setFitHeight(188);
-        preview.setPreserveRatio(false);
-        right.getChildren().add(preview);
-
-        main.setSpacing(20);
-        main.getChildren().addAll(left, right);
-
-        /* under header: review section*/
-        VBox reviews = new VBox();
-        HBox reviewHeading = new HBox();
-
-        Label reviewTitle = new Label("Ratings");
-        reviewTitle.getStyleClass().add("label3");
-        reviewTitle.getStyleClass().add("error");
-
-        double avgRating = new RatingService(new RatingRepository()).getAverageRating(s);
-        Text avgRatingText = new Text(avgRating > 0 ? String.format("(%.1f)", avgRating) : "(No ratings yet)");
-
-        avgRatingText.setStyle("-fx-font-size: 1.2em;");
-        reviewHeading.getChildren().addAll(reviewTitle, Stars.StarRow(avgRating, 1.2, 5), avgRatingText);
-        reviewHeading.setSpacing(10);
-        reviewHeading.setAlignment(Pos.CENTER_LEFT);
-
-        reviews.getChildren().addAll(reviewHeading);
-
-        base.getChildren().addAll(main, reviews);
-        ScrollPane wrapper = new ScrollPane(base);
-        wrapper.setFitToHeight(true);
-        wrapper.setFitToWidth(true);
-        instance.current.setCenter(wrapper);
+        MaterialPage.setPage(s);
     }
 
     public void displayProfile(int id) throws IOException {
@@ -247,53 +107,7 @@ public class SceneManager {
             return;
         }
 
-        VBox base = new VBox();
-        base.getStylesheets().add(Objects.requireNonNull(getClass().getResource("/css/style.css")).toExternalForm());
-        base.setSpacing(12);
-        base.setPadding(new Insets(20, 20, 20, 20));
-
-        Label title = new Label(u.getFullName());
-        title.getStyleClass().add("label3");
-        title.getStyleClass().add("error");
-        title.setWrapText(true);
-        title.setMaxWidth(600);
-
-        base.getChildren().addAll(title, TextLabels.getUserRoleLabel(u));
-
-        CategoryRepository cRepo = new CategoryRepository();
-        List<Category> userCategories = cRepo.findCategoriesByUser(u);
-
-        if (!userCategories.isEmpty()) {
-            Label cTitle = new Label("Courses");
-            cTitle.getStyleClass().add("label4");
-            cTitle.getStyleClass().add("secondary");
-
-            List<Category> categoryList = cRepo.findCategoriesByUser(u);
-            List<Button> buttonList = new ArrayList<>();
-
-            categoryList.forEach(c -> buttonList.add(listItemGraphic(c)));
-            ListView<Button> cView = ListItem.toListView(buttonList);
-
-            base.getChildren().addAll(cTitle, cView);
-        }
-
-        StudyMaterialRepository sRepo = new StudyMaterialRepository();
-        List<StudyMaterial> userMaterials = sRepo.findByUser(u);
-
-        if (!userMaterials.isEmpty()) {
-            Label mTitle = new Label("Materials");
-            mTitle.getStyleClass().add("label4");
-            mTitle.getStyleClass().add("primary-light");
-
-            List<Button> buttonList = new ArrayList<>();
-
-            userMaterials.forEach(m -> buttonList.add(listItemGraphic(m)));
-            ListView<Button> mView = ListItem.toListView(buttonList);
-
-            base.getChildren().addAll(mTitle, mView);
-        }
-
-        instance.current.setCenter(base);
+        ProfilePage.setPage(u);
     }
 
     public void displayErrorPage(String errorText, Screen redirectScreen, String redirectLabel) {
@@ -351,6 +165,10 @@ public class SceneManager {
         primaryStage.setTitle("StudyShelf");
         primaryStage.setScene(new Scene(instance.current, 800, 600));
         primaryStage.show();
+    }
+
+    public void setCenter(Node node){
+        instance.current.setCenter(node);
     }
 
     public void setPrimaryStage(Stage primaryStage){
